+++ src/main/java/com/twilio/rest/api/v2010/account/usage/record/AllTime.java
@@ -196,6 +196,7 @@ public class AllTime extends Resource {
         SMS_MESSAGES_FEATURES("sms-messages-features"),
         SMS_MESSAGES_FEATURES_SENDERID("sms-messages-features-senderid"),
         SPEECH_RECOGNITION("speech-recognition"),
+        STUDIO_ENGAGEMENTS("studio-engagements"),
         SYNC("sync"),
         SYNC_ACTIONS("sync-actions"),
         SYNC_ENDPOINT_HOURS("sync-endpoint-hours"),